Rename Footer folder to Regulatory
