+++ src/pages/Regulatory/CookiePolicy.jsx
