+++ src/pages/Regulatory/PrivacyNotice.jsx
