+++ src/pages/Regulatory/__tests__/CookiePolicy.test.jsx
