+++ src/pages/Regulatory/__tests__/PrivacyNotice.test.jsx
@@ -1,7 +1,7 @@
 import { render, screen } from '@testing-library/react';
 import PrivacyNotice from '../PrivacyNotice';
 
-describe('Cookie policy tests', () => {
+describe('Privacy notice tests', () => {
 
   it('should render a title of Privacy notice', async () => {
     render(<PrivacyNotice />);
